requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-auth==1.14.0
requests==2.23.0
pyrebase4==4.3.0
flask-api==2.0